requirements.txt
# OsMEN Python Dependencies

# Core dependencies
python-dotenv==1.0.0

# Qdrant client for vector database
qdrant-client==1.7.0

# HTTP requests
requests==2.31.0

# Data processing
pandas==2.1.4

# Date/time utilities
python-dateutil==2.8.2

# JSON utilities
jsonschema==4.20.0

# Subprocess and system interaction
psutil==5.9.6

# Logging
loguru==0.7.2

# Resilience and retry logic
tenacity==8.2.3

# Token encryption and security
cryptography==41.0.7

# RSS feed parsing for innovation monitoring
feedparser==6.0.10

# Document parsing for syllabus extraction
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Calendar integrations
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Microsoft integrations (for Outlook, To-Do, etc.)
msal==1.25.0

# Health data integrations (v1.6.0)
# android-health-connect - For Android Health Connect API (Android only)
# Note: android-health-connect requires Android environment

# Push notifications (v1.6.0)
firebase-admin==6.3.0

# Web Dashboard (v1.7.0)
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
itsdangerous==2.1.2
bcrypt==4.1.1
aiofiles==23.2.1
asyncpg==0.29.0
redis==5.0.1
prometheus-client==0.20.0
sentry-sdk==1.40.3

# Daily digest PDF export (v1.7.0)
reportlab==4.0.7

# Knowledge and task integrations (v1.8.0)
notion-client==2.2.1
todoist-api-python==2.1.5
watchdog==3.0.0
python-frontmatter==1.0.1

# Testing dependencies (Team 4)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.0
pytest-timeout>=2.1.0
pytest-html>=3.2.0
pytest-benchmark>=4.0.0
coverage>=7.3.0
responses>=0.23.0

# Optional: For enhanced functionality
# numpy==1.26.2
# langchain==0.1.0
# chromadb==0.4.18
ratelimit>=2.2.1

# DeepAgents integration (v3.0)
deepagents>=0.1.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
tavily-python>=0.5.0
numpy>=1.26.0

# LLM Provider Adapters (v3.0)
aiohttp>=3.9.0
openai>=1.0.0
anthropic>=0.30.0

# Librarian RAG Integration (optional - for local development)
# Note: Full librarian dependencies are in agents/librarian/requirements.txt
# These are minimal deps for using librarian_agent locally
chromadb>=0.5.0
sentence-transformers>=2.0.0
langgraph>=0.2.0

# ============================================================================
# OsMEN Extended Integrations (v3.1)
# ============================================================================

# System Monitoring
nvidia-ml-py>=12.0.0          # GPU monitoring (replaces deprecated pynvml)
pywin32>=306                   # Windows Task Scheduler, Registry access

# Voice/Audio Integration
audiblez>=0.4.9               # Audiobook creator (Kokoro TTS, EPUB/PDF to M4B)
kokoro-onnx>=0.3.3            # Kokoro TTS ONNX runtime (lightweight, 54 voices)
ebooklib>=0.18                # EPUB parsing
edge-tts>=6.0.0               # Microsoft Edge TTS (free, high quality)
pyttsx3>=2.90                 # Local TTS engine
soundfile>=0.12.0             # Audio file I/O
# faster-whisper>=0.10.0      # Local STT (optional - requires CUDA for best perf)
# elevenlabs>=0.2.0           # Premium TTS (optional - requires API key)

# Voice Cloning (optional - heavy dependencies, requires GPU for good performance)
# TTS>=0.22.0                 # Coqui TTS with XTTS for zero-shot voice cloning
# Note: Install manually with `pip install TTS` for voice cloning support

# Media Entertainment
# tmdb>=0.1.0                 # TMDB API (deprecated, using direct requests)
# Note: TMDB and OpenSubtitles use aiohttp (already included above)

# Creative Tools (optional - heavy dependencies)
# shap-e>=0.1.0               # Text-to-3D (optional - requires PyTorch)
# Note: InvokeAI and ComfyUI are separate installations

